backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi-mail>=1.5.8",
    "fastapi[standard]>=0.121.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.0.0",
    "langgraph>=1.0.3",
    "loguru>=0.7.3",
    "passlib[argon2]>=1.7.4",
    "pinecone[asyncio]>=7.3.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "sib-api-v3-sdk>=7.6.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
